Use explicit type for 'length' variable.
Replaced 'uint' with 'uint256' for clarity and consistency. This change aligns with best practices and ensures the explicit use of the full 256-bit unsigned integer type.

+++ contracts/ERC721Factory.sol
@@ -265,7 +265,7 @@ contract ERC721Factory is Ownable, Deployer, IERC721Factory {
         uri = string.concat(uri, '"', _publishData.name, '" : ');
         uri = string.concat(uri, _publishData.tokenURI, ', ');
 
-        uint length = accumulatederc20.length;
+        uint256 length = accumulatederc20.length;
 
         for (uint256 i = 0; i < length; i++) {
             dtaddress = accumulatederc20[i];
